Updated XercesUtils to solve problem in Visual Studio.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@1343 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/loaders/xml/XMLBasisTypes.cpp
@@ -451,7 +451,7 @@ float XMLBasisTypes::readFloat(xercesc::DOMElement* element, bool doCheckHeader)
     if (doCheckHeader)
         checkHeader(element, FloatId);
 
-    return std::atof(readElementText(element).c_str());
+    return (float)std::atof(readElementText(element).c_str());
 }
 
 int XMLBasisTypes::readInt(xercesc::DOMElement* element, bool doCheckHeader) {
